Implemented json configuration. Added example config;

+++ main.py
@@ -1,7 +1,10 @@
 from pynput.mouse import Button, Listener
-import subprocess, os, signal
+import subprocess, os, sys, signal, json
 
-#todo: debug
+#todo: replace with user home dir
+with open('./macros.json') as file:
+    macro_data = json.load(file)
+
 def print_button(x, y, button, pressed):
     action = "clicked" if pressed else "released"
     print("Mouse button {} {} at {} {}".format(button, action, x, y))
@@ -9,37 +12,44 @@ def print_button(x, y, button, pressed):
 process_list = {}
 
 def parse_mouse_action(x,y,button,pressed):
+    #skip left, right, middle entirely
+    if button == Button.left or button == Button.right or button == Button.middle:
+        return
+
     #on release of the button
-    if not pressed:
-        #todo: make dynamic, based on config
-        if button == Button.button13:
-            identifier = str(Button.button13)
-            if identifier in process_list:
-                pid = process_list[identifier]
+    if not pressed and str(button) in macro_data:
+        button_str = str(button)
+        macro = macro_data[button_str]
+        exec_file = "./{}_{}.py".format(macro['target_type'], macro['macro_type'])
+
+        if not os.path.isfile(exec_file) or not check_macro(macro):
+            print("invalid macro")
+            return
+
+        if macro['macro_type'] == 'spam':
+            if button_str in process_list:
+                pid = process_list[button_str]
                 os.killpg(os.getpgid(pid), signal.SIGTERM)
-                del(process_list[identifier])
+                del(process_list[button_str])
             else:
                 proc = subprocess.Popen(
-                    "./.venv/bin/python3 mouse_spammer.py left", #todo: fix
+                    "./.venv/bin/python3 {} {} {}".format(exec_file, macro['target'], macro['delay']),
                     shell=True,
                     preexec_fn=os.setsid
                 )
-                process_list[identifier] = proc.pid
-        elif button == Button.button12:
-            identifier = str(Button.button13)
-            if identifier in process_list:
-                pid = process_list[identifier]
-                os.killpg(os.getpgid(pid), signal.SIGTERM)
-                del(process_list[identifier])
-            else:
-                proc = subprocess.Popen(
-                    "./.venv/bin/python3 ./keyboard_spammer.py alt", #todo: fix
-                    shell=True,
-                    preexec_fn=os.setsid
-                )
-                process_list[identifier] = proc.pid
-        # else: #todo: for debugging
-        #     print(button)
+                process_list[button_str] = proc.pid
+        else:
+            print('others') #todo: implement other types of macros
+
+def check_macro(macro_obj):
+    if (
+            'macro_type' not in macro_obj or
+            'target_type' not in macro_obj or
+            'target' not in macro_obj or
+            'delay' not in macro_obj
+        ):
+        return False
+    return True
 
 with Listener(
     on_click=parse_mouse_action
